Split test_label_importers.py into 3 focused files
The original 1086-line file was too large to read in one pass. Split into:
- test_label_importers.py (~330 lines): base classes, registry, built-in importers
- test_label_import_endpoint.py (~310 lines): import endpoint, resolution helpers
- test_label_import_ingestion.py (~210 lines): ingestion endpoint and unit tests

Updated conftest.py test groups and CLAUDE.md documentation.
